+++ src/init/consentless.ts
@@ -3,9 +3,9 @@ import { initArticleInline } from 'init/consentless/dynamic/article-inline';
 import { initExclusionSlot } from 'init/consentless/dynamic/exclusion-slot';
 import { initFixedSlots } from 'init/consentless/init-fixed-slots';
 import { initConsentless } from 'init/consentless/prepare-ootag';
+import { reloadPageOnConsentChange } from 'init/shared/reload-page-on-consent-change';
 import { init as setAdTestCookie } from 'init/shared/set-adtest-cookie';
 import { init as setAdTestInLabelsCookie } from 'init/shared/set-adtest-in-labels-cookie';
-import { reloadPageOnConsentChange } from 'lib/reload-page-on-consent-change';
 
 const bootConsentless = async (consentState: ConsentState): Promise<void> => {
 	const consentlessModuleList = [
